Limit GLThread affinity for better graphic performance
BZ: 220429

Set the cpu affinity of the GLThread for the execution of onDrawFrame
and eglSwapBuffers functions. In case of core task switch there may be a
frequency penalty that results in missing frames and a UI jerk.
The old cpu affinity of the GThread is backuped so that if onDrawFrame
spawns threads or forks then the backuped affinity is set to the new threads.
If onDrawFrame changes the cpu affinity the backuped value is reseted
and no cpu affinity is restored when eglSwapBuffers completes.

Category: aosp improvement
Domain: AOSP-Framework-CoreService
Origin: internal
Upstream-Candidate: yes
Signed-off-by: Thomas Buhot <thomas.buhot@intel.com>

Conflicts:
	core/java/android/os/Process.java
	core/jni/android_util_Process.cpp

Change-Id: Id11919659e75bb18f5133c8901ecbda99421925e
Signed-off-by: charles.chen <cch@rock-chips.com>