Make automatic brightness control depend on FEATURE_SENSOR_LIGHT
The system defines several hardware/product build values in
frameworks/base/core/res/res/values/config.xml which are dynamically
checked in frameworks to provide the appropriate experience and
functionality.

In order to provide a single binary OS image that can operate on
devices that may or may not have an an ambient light sensor (ALS,)
changes are required to the framework/base where the build-time
resource config_automatic_brightness_control is used.

This patch changes any code which references
config_automatic_brightness_control or
def_screen_brightness_automatic_mode to first check if the platform
has the system feature PackageManager.FEATURE_SENSOR_LIGHT. If it
doesn't, those resource values are set to false.

A similar change is needed in the packages/apps/Settings project.

Change-Id: I79051d6a4d2c25ff6c099cc50df5e9ddd693e119
Signed-off-by: James Ketrenos <james.p.ketrenos@intel.com>
Reviewed-on: https://android.intel.com/316770
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Ortiz, Samuel <samuel.ortiz@intel.com>
Tested-by: Ortiz, Samuel <samuel.ortiz@intel.com>
Reviewed-by: Mathew, Alwin <alwin.mathew@intel.com>
Reviewed-on: https://android.intel.com/322655
Reviewed-by: Bhat, Sudheendra S <sudheendra.s.bhat@intel.com>
Tested-by: Grape, Markus <markus.grape@intel.com>
Reviewed-by: Fagerstedt, Axel <axel.fagerstedt@intel.com>
Signed-off-by: Cody Xie <xkd@rock-chips.com>