DimLayer : remove unnecessary surface transaction calls
There is the case that adjustSurface()  get called even if
the size of the surface of DimLayer is not changed actually.
Since changing the size of a surface is processed synchronously
in the SurfacFlinger, there is usually a few milliseconds delay
(up to 1 vsync interval) when we launch an application.
This patch avoids such cases.

Change-Id: Ib1f76d54f9f2364ac54b70120e4b781e8534e750
Signed-off-by: Dohyun Lee <dohyun.lee@lge.com>
(cherry picked from commit 6ce3a37f6b06585c5c0c045d1df69f21f8200067)

Signed-off-by: Cody Xie <xkd@rock-chips.com>