PowerProfile: Read power values from an xml file first
This change modifies the PowerProfile class so it first checks and tries
to parse the power values from the file /system/etc/power_profile.xml.
If the file does not exist or generates a parsing exceptios then it
fallbacks to the legacy xml resource.

This framework modification is approved by Google and tracked under
buganizer 19460207.

Issue: AREQ-5180
Change-Id: I48d67863e72315b49927599c41a7c411ecb847ef
Signed-off-by: Thierry Escande <thierry.escande@linux.intel.com>
Reviewed-on: https://android.intel.com/349565
Reviewed-by: Keel, Thomas <thomas.keel@intel.com>
Reviewed-by: Chen, Hu1 <hu1.chen@intel.com>
Reviewed-by: Escande, Thierry <thierry.escande@intel.com>
Tested-by: Escande, Thierry <thierry.escande@intel.com>
Signed-off-by: Cody Xie <xkd@rock-chips.com>