ConnectivityService: fix mms issue when data disabled
We support the feature that mms can be sent out even if mobile
data is disabled. But after mms transaction is completed, the
data connection is not disconnected. Because the new NetworkAgent
which be created for mms satifies default mobile data Request also
, and want to be kept.

Solution: Check if default mobile data Request should be added to
new NetworkAgent.
1. networkType is ConnectivityManager.TYPE_MOBILE
2. NetworkAgent's connection comes from data SIM
3. mobile data is enabled