Optimze boot up register NW time interval
StatusBar often take for about 10s to remove 'no SIM' icon
and display 'signal strength' icon while boot up with SIMs.
Then if NW is already camped, it have to wait status bar
to update the icons.
To reduce the time inerval, check hasIccCard too if subInfo
is not ready, it should be much earlier.