SystemUI: fix the signal bar miss sometimes
After turn on/off the sim, sometimes signal bar miss.
need to click the status bar to update.
Because after SUB list changed, mobile controller clear all
signal bar, and new the signal bar for the inserted SIM.
But the signal bar already in the cache, so it just get from
cache, but the state is same to last, so it doesn't call
notifyListener to draw it.

Solution: After get the signal bar from cache, notify listener
update it.